Function unique_mesh_name() moved in a new file (utilities.py)
